1.添加腾讯云COS信息，完成图像上传功能 2.修正提示逻辑 3.修正Comment类中错误映射 4.修正CommentController中RequestMapping配置错误 5.添加评论时文章存在的判断方法 6.完善删除评论时传参错误以及权限判断 7.改良评论查询，使评论查询适应更多环境 8.添加管理员审核评论的方法 9.添加通过评论id查找评论实体的便捷方法 10.改良评论查询Mapper，贴合第7点 11.修复上传文件时文件为空的bug 12.测试用例中改良加密方式，使得加密/解密更方便

+++ src/main/java/com/bilicute/spacetime/controller/CommentController.java
@@ -8,17 +8,22 @@ import com.bilicute.spacetime.quickMethods.QuickMethods;
 import com.bilicute.spacetime.service.ArticleService;
 import com.bilicute.spacetime.service.CommentService;
 import jakarta.validation.constraints.Min;
+import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 import java.util.Map;
+import java.util.Objects;
 
 
-@RestController("/comment")
+@RestController
+@RequestMapping("/comment")
 @Slf4j
+@Validated
 public class CommentController {
     private static ArticleService articleService;
     @Autowired
@@ -31,12 +36,20 @@ public class CommentController {
 
     @PostMapping("/add")
     public Result addComment(String content,Integer articleId) {
+        Article article = articleService.findById(articleId);
+        if (article == null) {
+            return Result.error("所属文章不存在");
+        }
         commentService.add(content,articleId);
         return Result.success();
     }
 
     @DeleteMapping("/delete")
-    public Result deleteComment(Integer commentId) {
+    public Result deleteComment(@NotNull(message = "请输入待删除的评论") Integer commentId) {
+        Comment theComment = commentService.findById(commentId);
+        if (!QuickMethods.isAdmin()|| !Objects.equals(theComment.getCreateUser(), QuickMethods.getLoggedInUserId())){
+            return Result.error("没有权限删除");
+        }
         commentService.delete(commentId);
         return Result.success();
     }
@@ -51,19 +64,27 @@ public class CommentController {
         return Result.success(number);
     }
 
-    @GetMapping
+    @GetMapping("/get")
     public Result<PageBean<Comment>> list(
             //当前页码
             Integer pageNum,
             //每页条数
             Integer pageSize,
             //文章ID
-            Integer articleId,
+            @RequestParam(required = false) Integer articleId,
             //审核状态（可选）
             @RequestParam(required = false) Boolean auditingState
     ){
+//        if (!(articleId != null&&articleService.findById(articleId)!=null)) {
+//            PageBean<Comment> pageBean = new PageBean<Comment>();
+//            Comment comment = new Comment();
+//            comment.setContent("指定文章不存在");
+//            List<Comment> list = List.of(new Comment[]{comment});
+//            pageBean.setItems(list);
+//            return Result.errorData(pageBean);
+//        }
         //如果获取审核的列表
-        if (!auditingState){
+        if (auditingState!=null&&!auditingState){
             //如果不是管理员
             if (!QuickMethods.isAdmin()){
                 PageBean<Comment> pageBean = new PageBean<Comment>();
@@ -79,6 +100,22 @@ public class CommentController {
         }
         PageBean<Comment> commentPageBean = commentService.list(pageNum,pageSize,articleId,auditingState);
         return Result.success(commentPageBean);
+    }
+
+    @PatchMapping("/checked")
+    public Result<?> checked(@NotNull(message = "未指定评论id") Integer id){
+        Comment comment = commentService.findById(id);
+        if (comment == null) {
+            return Result.error("指定评论不存在");
+        }
+        if(!QuickMethods.isAdmin()){
+            return Result.error("权限不足");
+        }
+        if (comment.getAuditingState()){
+            return Result.error("该评论已通过审核，无需再次审核");
+        }
+        commentService.check(id);
+        return Result.success();
     }
 
 }
